+++ Com.Danliris.Service.Inventory.Mongo.Lib/MongoModels/InventorySummaryMongo.cs
@@ -19,13 +19,13 @@ namespace Com.Danliris.Service.Inventory.Mongo.Lib.MongoModels
 
         public string storageName { get; set; }
 
-        public int quantity { get; set; }
+        public double quantity { get; set; }
 
         public ObjectId uomId { get; set; }
 
         public string uom { get; set; }
 
-        public int stockPlanning { get; set; }
+        public double stockPlanning { get; set; }
 
         public string code { get; set; }
     }
